+++ src/main/scala/tscfg/Struct.scala
@@ -5,6 +5,7 @@ import tscfg.DefineCase.{
   AbstractDefineCase,
   EnumDefineCase,
   ExtendsDefineCase,
+  ImplementsDefineCase,
   SimpleDefineCase
 }
 import tscfg.exceptions.ObjectDefinitionException
@@ -31,8 +32,9 @@ case class Struct(
   def isDefine: Boolean = defineCaseOpt.isDefined
 
   def isExtends: Boolean = defineCaseOpt match {
-    case Some(_: ExtendsDefineCase) => true
-    case _                          => false
+    case Some(_: ExtendsDefineCase)    => true
+    case Some(_: ImplementsDefineCase) => true
+    case _                             => false
   }
 
   def isEnum: Boolean = defineCaseOpt.exists(_.isEnum)
@@ -98,12 +100,8 @@ object Struct {
     def addExtendStruct(
         s: Struct,
         othersBeingAdded: List[Struct] = List.empty
-    ): Unit = s.defineCaseOpt.get match {
-
-      case SimpleDefineCase | AbstractDefineCase | EnumDefineCase =>
-        sorted.put(s.name, s)
-
-      case ExtendsDefineCase(name, _) =>
+    ): Unit = {
+      def addExtendsOrImplements(name: String): Unit = {
         sorted.get(name) match {
           case Some(_) =>
             // ancestor already added. Add this descendent struct:
@@ -117,7 +115,9 @@ object Struct {
                 if (othersBeingAdded.exists(_.name == ancestor.name)) {
                   val path = s :: othersBeingAdded
                   val via =
-                    path.reverse.map("'" + _.name + "'").mkString(" -> ")
+                    path.reverseIterator
+                      .map("'" + _.name + "'")
+                      .mkString(" -> ")
                   throw ObjectDefinitionException(
                     s"extension of struct '${s.name}' involves circular reference via $via"
                   )
@@ -137,7 +137,20 @@ object Struct {
                 )
             }
         }
+      }
+
+      s.defineCaseOpt.get match {
+        case SimpleDefineCase | AbstractDefineCase | EnumDefineCase =>
+          sorted.put(s.name, s)
+
+        case ExtendsDefineCase(name, _) =>
+          addExtendsOrImplements(name)
+
+        case ImplementsDefineCase(name, _) =>
+          addExtendsOrImplements(name)
+      }
     }
+
     defineStructs.foreach(addExtendStruct(_))
 
     assert(
@@ -167,39 +180,50 @@ object Struct {
       memberStructs: List[Struct],
       namespace: Namespace
   ): Option[Map[String, model.AnnType]] = {
+
+    def handleExtends(
+        parentName: String
+    ): Option[Map[String, model.AnnType]] = {
+      val defineStructs = memberStructs.filter(_.isDefine)
+
+      val greatAncestorMembers =
+        defineStructs.find(_.name == parentName) match {
+          case Some(parentStruct) if parentStruct.isExtends =>
+            ancestorClassMembers(parentStruct, memberStructs, namespace)
+
+          case Some(_) => None
+
+          case None if isExternalParent(parentName) => None
+
+          case None =>
+            throw new RuntimeException(
+              s"struct '${struct.name}' with undefined extend '$parentName'"
+            )
+        }
+
+      val parentMembers =
+        namespace.getRealDefine(parentName).map(_.members) match {
+          case Some(parentMembers) => parentMembers
+
+          case None if isExternalParent(parentName) => None
+
+          case None =>
+            throw new IllegalArgumentException(
+              s"@define '${struct.name}' is invalid because '$parentName' is not @defined"
+            )
+        }
+
+      // join both member maps
+      Some(greatAncestorMembers.getOrElse(Map.empty) ++ parentMembers)
+    }
+
     struct.defineCaseOpt.flatMap {
       case ExtendsDefineCase(parentName, _) =>
-        val defineStructs = memberStructs.filter(_.isDefine)
+        handleExtends(parentName)
 
-        val greatAncestorMembers =
-          defineStructs.find(_.name == parentName) match {
-            case Some(parentStruct) if parentStruct.isExtends =>
-              ancestorClassMembers(parentStruct, memberStructs, namespace)
-
-            case Some(_) => None
-
-            case None if isExternalParent(parentName) => None
-
-            case None =>
-              throw new RuntimeException(
-                s"struct '${struct.name}' with undefined extend '$parentName'"
-              )
-          }
-
-        val parentMembers =
-          namespace.getRealDefine(parentName).map(_.members) match {
-            case s @ Some(parentMembers) => parentMembers
-
-            case None if isExternalParent(parentName) => None
-
-            case None =>
-              throw new IllegalArgumentException(
-                s"@define '${struct.name}' is invalid because '$parentName' is not @defined"
-              )
-          }
-
-        // join both member maps
-        Some(greatAncestorMembers.getOrElse(Map.empty) ++ parentMembers)
+      case ImplementsDefineCase(parentName, _) =>
+        // note: handling it as an extends (todo: revisit this at some point)
+        handleExtends(parentName)
 
       case _ => None
     }